Calculation of the voter turnout difference between two vote topics ("Vorlagen")

Calculation of the voter turnout difference between two vote topics ("Vorlagen")

Usage

cross_fun(df, vorl1, vorl2, geo_cols = geocols)

Arguments

df

A data frame containing the municipality ID and voter turnout for various proposals. Each column represents a specific proposal, and the column names should correspond to the proposal IDs (e.g., 'eidg1', 'kant2').

vorl1

A character vector specifying the proposal ID for the first proposal (e.g., "eidg1", "kant2"). The proposal ID should match the column name in the dataset.

vorl2

A character vector specifying the proposal ID for the second proposal (e.g., "eidg1", "kant2"). This should also correspond to a column name in the dataset.

geocols

Geo-Column This should also correspond to a column name in the dataset.

Value

Tibble with the turnout difference between two vote topics

Examples

testdata <- tibble(gemwkid = c(13,49,41,43,44),
eidg1 = c(60.90,61.18,65.27,55.36,57.68),
eidg2 = c(62.16,62.54,66.95,56.65,58.68),
kant1 = c(57.73,60.27,63.31,51.93,54.49))

cross_fun(testdata,"eidg1","eidg2")
#> Error in dplyr::select(., tidyselect::all_of(geo_cols), !!varname): ℹ In argument: `tidyselect::all_of(geo_cols)`.
#> Caused by error in `cross_fun()`:
#> ! object 'geocols' not found

 # generate combinations
combinations <-as.data.frame(t(combn(c("eidg1","eidg2","kant1"),2)))

# difference between columns named as the first combination
cross_fun(crosscheckdata,combinations$V1[1],combinations$V2[1])
#> Error: object 'crosscheckdata' not found