Calculation of voter turnout differences between all vote topics ("Vorlagen")

Calculation of voter turnout differences between all vote topics ("Vorlagen")

get_differences(df, vorl1, vorl2, geo_cols = c("gemwkid", "gemeinde"))

Arguments

df

A data frame containing the municipality ID and voter turnout for various vote topics, with each vote topic in a separate column.

vorl1

A vector of proposal IDs corresponding to column names in the dataset (e.g., "eidg1", "kant2").

vorl2

A vector of proposal IDs (e.g., "eidg1", "kant2") for comparison with `vorl1`.

Value

A tibble containing voter turnout differences between all possible combinations of vote topics.

Examples

library(tibble)

testdata <- tibble(gemwkid = c(13,49,41,43,44),
eidg1 = c(60.90,61.18,65.27,55.36,57.68),
eidg2 = c(62.16,62.54,66.95,56.65,58.68),
kant1 = c(57.73,60.27,63.31,51.93,54.49))

single_difference(testdata,"eidg1","eidg2")
#> Error in single_difference(testdata, "eidg1", "eidg2"): could not find function "single_difference"

 # generate combinations
combinations <-as.data.frame(t(combn(c("eidg1","eidg2","kant1"),2)))

# calculate all possible differences between columns

get_differences(testdata,combinations$V1,combinations$V2)
#> Error in map2(.x, .y, .f, ...): ℹ In index: 1.
#> Caused by error in `dplyr::select()`:
#> ℹ In argument: `tidyselect::all_of(geo_cols)`.
#> Caused by error in `tidyselect::all_of()`:
#> ! Can't subset elements that don't exist.
#> ✖ Element `gemeinde` doesn't exist.