Run prediction for one vote

This function can be used to predict the outcome of one vote based on a trained model, generated by using the plausi::train_prediction_model().

Usage

predict_single_vote(model, testdata)

Arguments

model

A trained model, generated using plausi::train_prediction_model().

testdata

Dataset on which the prediction should be run. The data must contain all columns of the training data of the model model$trainingData.

Value

A data.frame.

Examples

test_model <- train_prediction_model("Eidg1", votedata, to_exclude_vars = "Kant1")

predict_single_vote(test_model, votedata)
#> # A tibble: 171 × 5
#>    gemeinde           v_gemwkid  pred  real vorlage
#>    <chr>                  <dbl> <dbl> <dbl> <chr>  
#>  1 Adlikon                   21  23.2  21.5 Eidg1  
#>  2 Adliswil                 131  47.4  48.3 Eidg1  
#>  3 Aesch                    241  30.0  30.9 Eidg1  
#>  4 Aeugst am Albis            1  32.7  31.5 Eidg1  
#>  5 Affoltern am Albis         2  40.7  39.8 Eidg1  
#>  6 Altikon                  211  28.9  29.8 Eidg1  
#>  7 Andelfingen               30  33.0  32.1 Eidg1  
#>  8 Bachenbülach              51  38.6  39.9 Eidg1  
#>  9 Bachs                     81  31.4  30.5 Eidg1  
#> 10 Bäretswil                111  32.7  33.3 Eidg1  
#> # ℹ 161 more rows