Run predictions for multiple columns (specifically votes) in a dataset

Run predictions for multiple columns (specifically votes) in a dataset

predict_votes(
  votes,
  train,
  test = NULL,
  method = "bagEarth",
  trControl = NULL,
  exclude_votes = FALSE,
  geovars = c("gemeinde", "v_gemwkid"),
  testprop = NA,
  ...
)

Arguments

votes

names of the dependent variable-columns

train

data used to train the model containing the variables to be predicted and the predictor-columns

test

optional dataset structured identically as the trainingdataset on which the prediction

method

method available in the caret-package which should be used for the prediction

trControl

parameters to tune the model

exclude_votes

if TRUE the variables to be predicted will be excluded from each others models

geovars

variables containing labels and ids of the spatial units

testprop

optional argument to generate a training dataset by splitting the dataset (testprop=share of observations to be randomly kept)

...

optional parameters that can be passed to the caret::train function

Value

data.frame

Examples

predict_votes(c("Eidg1","Kant1"), votedata, exclude_votes=TRUE)
#> Error in map(.x, .f, ...): ℹ In index: 1.
#> Caused by error:
#> ! Required packages are missing: earth