Calculate RMSE

Calculate the Root Mean Square Error (RMSE). The RMSE is the standard deviation of the residuals (prediction errors) and therefore an indicator of how precise the prediction of a specific vote actually is.

Usage

rmse(prediction, observation, na.rm = TRUE)

Arguments

prediction

Predicted value.

observation

Oserved value.

na.rm

Remove NA values, defaults to TRUE

Value

A vector of numeric values.

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(tidyr)
library(kernlab)
#> 
#> Attaching package: ‘kernlab’
#> The following object is masked from ‘package:ggplot2’:
#> 
#>     alpha

pred_data  <- predict_multiple_votes(c("Eidg1","Kant1"), votedata, exclude_votes = TRUE)

pred_data %>%
drop_na() %>%
group_by(vorlage) %>%
summarize(rmse=RMSE(pred,real))
#> # A tibble: 2 × 2
#>   vorlage  rmse
#>   <chr>   <dbl>
#> 1 Eidg1    2.06
#> 2 Kant1    1.66